IM LOOOOSING IT..pum, pum, purum

+++ test.php
@@ -110,11 +110,6 @@ function TestFiles($source)
 			return false;
 		}
 
-		// if(!isXmlStructureValid($source) and $int_only_flag == true)	//zistuje ci .src je xml subor
-		// {
-		// 	return false;
-		// }
-
 		if(!file_exists($ref_rc))
 		{
 			$test_file = fopen($ref_rc, "w");
@@ -161,7 +156,10 @@ function TestFiles($source)
 		else
 		{
 			exec("php7.4 ".$parse_file ." <$source >temp_output", $result,$rc_out);
-			exec("python3.8 ".$int_file ." --source=temp_output <$ref_in >temp_output2", $result,$rc_out);
+			if ($rc_out == 0)
+			{ 
+				exec("python3.8 ".$int_file ." --source=temp_output <$ref_in >temp_output2", $result,$rc_out);
+			}
 		}
 		if($rc_out != $exp_rc)
 		{
@@ -200,8 +198,8 @@ function TestFiles($source)
 			}
 			
 		}
-		unlink("temp_output");
-		unlink("temp_output2");
+		if($int_only_flag == false) {unlink("temp_output");}
+		if($parse_only_flag == false) {unlink("temp_output2");}
 	} 
 	return True;	
 }
